refactor tc 개선
