+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.RacingGameFactory;
 import racingcar.domain.Car;
 import racingcar.domain.RacingCars;
 import racingcar.domain.RacingGame;
@@ -22,7 +23,6 @@ public class RacingGameTest {
         List<String> names = Arrays.asList("car1","car2","car3");
 
         RacingGame racingGame = new RacingGame(getRacingCars(names, canMove));
-
         racingGame.tryMove();
 
         racingGame.getCars()
@@ -34,7 +34,7 @@ public class RacingGameTest {
     @ValueSource(strings = {"car1,car2,car3"})
     void getPositions(String carNames) {
         List<String> names = Arrays.asList(carNames.split(","));
-        RacingGame racingGame = RacingGameFactory.newRacingGame(names);
+        RacingGame racingGame = createRacingGame(carNames);
 
         List<Car> cars = racingGame.getCars();
         for (int i = 0; i < names.size() ; i++){
@@ -49,16 +49,12 @@ public class RacingGameTest {
     @ParameterizedTest
     @ValueSource(strings = {"car1,car2,car3"})
     void getWinners(String carNames) {
-        List<String> names = Arrays.asList(carNames.split(","));
-        RacingGame racingGame = RacingGameFactory.newRacingGame(names);
+        RacingGame racingGame = createRacingGame(carNames);
         racingGame.tryMove();
-
 
         int max= racingGame.getCars().stream().map(Car::getPosition).max(Integer::compareTo).orElse(0);
 
         racingGame.getWinners().forEach(car -> assertThat(car.getPosition()).isEqualTo(max));
-
-
     }
 
     private RacingCars getRacingCars(List<String> names, boolean canMove) {
@@ -67,6 +63,11 @@ public class RacingGameTest {
                 .map(s -> new Car(() -> canMove))
                 .collect(Collectors.toList())
         );
+    }
+
+    private RacingGame createRacingGame(String carNames){
+        List<String> names = Arrays.asList(carNames.split(","));
+        return  RacingGameFactory.newRacingGame(names);
     }
