+++ src/test/java/racingcar/domain/RandomCarMovingConditionTest.java
@@ -1,11 +1,10 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.domain.RandomCarMovingCondition;
 
 class RandomCarMovingConditionTest {
